requirements.txt
fastapi
uvicorn
llama-cpp-python
pytest
pydantic